- cleaned up emailer_permission_test - pep 8 cleanup - fixed the test to check for correct text after emailer UI update - renamed files/ Camel cased classes
Further encapsulation can be done in the Emailer_Test (in a later commit)
- URL string
- assert into an is_authorized function
- PO web driver waits
